+++ lib/dimension.ts
@@ -1,7 +1,9 @@
 import { Block } from "./block.js";
 import { Entity } from "./entity.js";
+import { ItemStack } from "./item/stack.js";
 import { Wrapper } from "./wrapper.js";
 import { map } from "./iterable.js";
+import { Vector3 } from "@minecraft/server";
 import * as MC from "@minecraft/server";
 
 export class Dimension extends Wrapper<MC.Dimension> {
@@ -20,7 +22,11 @@ export class Dimension extends Wrapper<MC.Dimension> {
         });
     }
 
-    public spawnEntity(identifier: string, location: MC.Vector3): Entity {
+    public spawnEntity(identifier: string, location: Vector3): Entity {
         return new Entity(this.raw.spawnEntity(identifier, location));
+    }
+
+    public spawnItem(itemStack: ItemStack, location: Vector3): Entity {
+        return new Entity(this.raw.spawnItem(itemStack.raw, location));
     }
 }