+++ NEWS.md
@@ -4,7 +4,8 @@
 
 * New properties:
   * `Block.prototype.getItemStack`
-  * Setter for `Block.prototype.type`.
+  * `Block.prototype.type` setter
+  * `Dimension.prototype.spawnItem`
   * `World.prototype.getDimension`
 
 ## 2.0.0 -- 2023-09-17